add photo mode to talk photo

+++ src/CapstoneCodeV1/Control_Options_Terminal.py
@@ -192,7 +192,7 @@ def Terminal_Control(motor):
             [inch] = [float(element.strip()) for element in elements[1:]]
 
         if mode == 'photo':
-            number = [float(element.strip()) for element in elements[1:]]
+            number = [int(element.strip()) for element in elements[1:]]
             command = 'raspistill -o '+ number + '.jpg -h 640 -w 640 -t 10 -rot 0'
             print(command)
             os.system(command)
